[CELEBORN-179][BUG] Repeat remove expired shuffle throw NPE (#1124)

+++ worker/src/main/scala/org/apache/celeborn/service/deploy/worker/storage/StorageManager.scala
@@ -390,15 +390,17 @@ final private[worker] class StorageManager(conf: CelebornConf, workerSource: Abs
   def cleanupExpiredShuffleKey(expiredShuffleKeys: util.HashSet[String]): Unit = {
     expiredShuffleKeys.asScala.foreach { shuffleKey =>
       logInfo(s"Cleanup expired shuffle $shuffleKey.")
-      val hdfsInfos = fileInfos.remove(shuffleKey).asScala.filter(_._2.isHdfs)
-      val (appId, shuffleId) = Utils.splitShuffleKey(shuffleKey)
-      disksSnapshot().filter(_.status != DiskStatus.IO_HANG).foreach { diskInfo =>
-        diskInfo.dirs.foreach { dir =>
-          val file = new File(dir, s"$appId/$shuffleId")
-          deleteDirectory(file, diskOperators.get(diskInfo.mountPoint))
+      if (fileInfos.containsKey(shuffleKey)) {
+        val hdfsInfos = fileInfos.remove(shuffleKey).asScala.filter(_._2.isHdfs)
+        val (appId, shuffleId) = Utils.splitShuffleKey(shuffleKey)
+        disksSnapshot().filter(_.status != DiskStatus.IO_HANG).foreach { diskInfo =>
+          diskInfo.dirs.foreach { dir =>
+            val file = new File(dir, s"$appId/$shuffleId")
+            deleteDirectory(file, diskOperators.get(diskInfo.mountPoint))
+          }
         }
+        hdfsInfos.foreach(item => item._2.deleteAllFiles(StorageManager.hdfsFs))
       }
-      hdfsInfos.foreach(item => item._2.deleteAllFiles(StorageManager.hdfsFs))
     }
   }
